requirements.txt
numpy
matplotlib
pytest
pytorch
sphinx